requirements.txt
Flask==3.1.1
gunicorn==21.2.0

langchain==1.2.6
langchain-core==1.2.7
langchain-community==0.4.1
langchain-openai==1.1.7
langchain-huggingface==1.2.0
langchain-pinecone==0.2.13
langchain-text-splitters==1.1.0

pinecone==7.3.0
sentence-transformers==4.1.0
transformers==4.57.6
torch==2.9.1

openai==1.109.1
python-dotenv==1.1.0
requests==2.32.5
numpy==2.4.1
scikit-learn==1.8.0
